* Added DragScroll contrib plugin in build system. * Fixed all 64bit compatibility issues. * Fixed all AngelScript related issues in non-x86 platforms. Scripting isn't available inside C::B for those platforms, but it builds and runs correctly.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@2033 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/as/source/as_context.cpp
@@ -2607,7 +2607,7 @@ int asCContext::CallGeneric(int id, void *objectPointer)
 			}
 
 			// Add the base offset for multiple inheritance
-			currentObject = (void*)(int(currentObject) + sysFunc->baseOffset);
+			currentObject = (void*)(long(currentObject) + sysFunc->baseOffset);
 
 			// Keep a reference to the object to protect it 
 			// from being released before the method returns
